refactor : postPoint 타입 변경

+++ src/main/java/com/kakaotech/team14backend/outer/post/dto/GetPopularPostResponseDTO.java
@@ -2,5 +2,5 @@ package com.kakaotech.team14backend.outer.post.dto;
 
 import java.util.List;
 
-public record GetPopularPostResponseDTO(Long postId, String imageUrl, List<String> hashTags, Long likeCount, Integer postPoint, String nickname,Boolean isLiked, Integer postLevel) {
+public record GetPopularPostResponseDTO(Long postId, String imageUrl, List<String> hashTags, Long likeCount, Long postPoint, String nickname,Boolean isLiked, Integer postLevel) {
 }
